chore: atom과 selector 분리
